Implement better pagination handling, and integrate into application
Used the available methods in the zendesk ruby api gem to paginate the response when getting all tickets from an account from the Tickets class.
Adjusted the Application class to accomodate these changes, cleaning it up in the process.
Cleaned up the ticket class as well, removing extraneous comments, and adding a field for the ticket ID